backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
openai==1.68.2
pydantic==2.11.1
python-dotenv==1.1.0
httpx==0.28.1
aiosqlite==0.21.0
asyncpg==0.30.0
bcrypt==4.3.0
PyJWT==2.10.1
